Allow the primitives for GHC.ExecutionStack
There's still work left. In particular the Dwarf functions should call
dwarf_init() by themselves. This is to simplify the Haskell API later
on. Also dwarf_init() have to change, so that calling it twice has no
effect.

NOTE: 5 Feb 2014: I "fixuped" the delete commits later, and now
this should be an empty commit, but isnt due to the shit that
I committed aswell when I first commited