Add module ExecutionStack.hsc to base
The module is an Haskell interface to displaying Stack traces based on
the execution stack. The interface is meant to be simple enough to be
exposed directly to any intermediate Haskell user.

This new module relies on:

     * Peter Wortmann's DWARF generation patches (-g flag)
     * The section `.debug_ghc` section, which was *not* introduced in
       https://phabricator.haskell.org/D396
     * Dwarf.c and the structure it assumes on the debug_ghc section.
     * Some exposed rts functions in Dwarf.h.

Note that this patch does not contain the changes for retrieving the
stack trace port-mortem.  That is a different beast and more
controversial, as it would be more changes rather than only additions.

Conflicts:
	libraries/base/tests/all.T

Arash: Amend test 004 and Add test 005

Comments 005.hs