Haskell land Stack Traces proposal
Summary:
This is definetly not ready for merging. I mainly want to discuss the
Haskell interface we're exposing here.  I rebased @scpmw's `.debug_ghc` section
patch and his Dwarf reading stuff here, just so that the Haskell interface has
some actual working implementation to rely on. As Peter said in D396, he is not
happy with the way the section is generated now (is that correct Peter?).

The other thing is that I want to show that we're not *that* far from having
low overhead stack traces now that D169 and D396 are landed. I counted this
patch set to about 2500 lines of code.

Test Plan:
I've added test cases that test the whole chain (including Peter's whole `-g`
architecture) but some of the tests I added actually fail now. Will have to
check into that.

Reviewers: scpmw, simonmar, austin, tibbe, hvr

Subscribers: thomie, scpmw

Differential Revision: https://phabricator.haskell.org/D662

GHC Trac Issues: #3693