Add module ExecutionStack.hsc to base
The module is an Haskell interface to displaying Stack traces based on
the execution stack. The interface is meant to be simple enough to be
exposed directly to any intermediate Haskell user.

This new module relies on:

     * Peter Wortmann's DWARF generation patches (-g flag)
     * Some exposed rts functions in Dwarf.h.

Note that this patch does not contain the changes for retrieving the
stack trace port-mortem.