rts: Add Stack.c and accompanying rtsflags
This new module is meant to contain C-functions that operate on the
execution stack and its members (the stack frames).  My intent at the
moment is to come one step closer to implementing stack traces.

I wait with testing this module and the rtsflags. I thought of doing
full-stack (pun intended) tests by testing that stack-traces are
available from Haskell land in pure Haskell tests. In those tests, we'll
be using the rtsflags as well.

Also, the changes here does not affect existing code paths that much. It
mainly just adds new functions.