Add Stack Traces in Haskell land based on symtab
Summary:
This patch extends the base library with a new Haskell module called
GHC.ExecutionStack. Essentially it provides the simple function
"printCurrentExecutionStack", which will print a stack trace to stdout.

As the module name suggests, it uses GHC-specific implementation details
to retrieve a stack. In particular it uses the stg stack.

The usefulness of the printed stack is currently low. As it's only using
names from the symbol table, which most of the time are just random
(something like "s2cZ_info"). But you can sometimes get useful names in
the stack trace still (like "Main_myfunction_info"), so I it is still
useful when debugging. See executionStack001.hs for an example when it
could print something useful.

Test Plan:
Run validate script. I also added 4 test cases.

Note that this patch doesn't change any existing code paths, so it's quite
unintrusive in that regard.

Reviewers: austin, hvr

Subscribers: bgamari, thomie

Differential Revision: https://phabricator.haskell.org/D963

GHC Trac Issues: #3693